+++ src/components/notifications/NotificationListItem.jsx
@@ -23,11 +23,15 @@ export default function NotificationListItem({
 
         <Container column className={style.content} gap={12}>
           <Container>
-            <Text fontSize={12}>{content}</Text>
+            <Text fontSize={{ base: 12, md: 10, '2xl': 12 }}>{content}</Text>
           </Container>
 
           <Container>
-            <Text fontSize={12} color="gray.650" lineHeight={1}>
+            <Text
+              fontSize={{ base: 12, md: 10, '2xl': 12 }}
+              color="gray.650"
+              lineHeight={1}
+            >
               {DateTime.fromISO(create_date).toRelative()}
             </Text>
           </Container>