+++ src/slices/AppSlice.js
@@ -2,10 +2,11 @@ import { createSlice } from '@reduxjs/toolkit'
 import { uuid4 } from '@utils'
 
 export const AppReducer = createSlice({
-  name: 'friends',
+  name: 'app',
   initialState: {
     toasts: [],
     friendListOpen: false,
+    maintenance: false,
   },
   reducers: {
     addToast: (state, action) => {
@@ -23,9 +24,17 @@ export const AppReducer = createSlice({
     toggleFriendList: (state, action) => {
       return { ...state, friendListOpen: action.payload }
     },
+
+    updateMaintenance: (state, action) => {
+      return {
+        ...state,
+        maintenance: action.payload,
+      }
+    },
   },
 })
 
-export const { addToast, removeToast, toggleFriendList } = AppReducer.actions
+export const { addToast, removeToast, toggleFriendList, updateMaintenance } =
+  AppReducer.actions
 
 export default AppReducer.reducer