+++ src/views/lobby/LobbyView.jsx
@@ -1,4 +1,4 @@
-import { Badge, Icon, Text } from '@chakra-ui/react'
+import { Badge, Icon, Text, useMediaQuery } from '@chakra-ui/react'
 import { useEffect, useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 
@@ -8,7 +8,6 @@ import {
   Container,
   JoystickIcon,
   LobbyLineup,
-  LobbyPlayButton,
   MatchFoundModal,
 } from '@components'
 import { StorageService } from '@services'
@@ -18,9 +17,10 @@ import { removeRestartQueue } from '@slices/LobbySlice'
 import style from './LobbyView.module.css'
 
 export default function LobbyView() {
+  const [isLessThan2xl] = useMediaQuery('(max-width: 1600px)')
+
   const user = useSelector((state) => state.user)
   const lobby = useSelector((state) => state.lobby)
-
   const preMatch = useSelector((state) => state.preMatch)
   const match = useSelector((state) => state.match)
 
@@ -28,39 +28,10 @@ export default function LobbyView() {
 
   const [openMatchFoundModal, setOpenMatchFoundModal] = useState(false)
 
-  const isOwner = lobby.owner_id === user.id
   const userPlayer = lobby.players?.find((player) => player.user_id === user.id)
   const otherPlayers = lobby.players?.filter(
     (player) => player.user_id !== user.id
   )
-
-  const handleQueue = async (action) => {
-    if (preMatch || match) return
-
-    if (action === 'start' && !isOwner) return
-
-    const userToken = StorageService.get('token')
-
-    let response = null
-
-    if (action === 'start') {
-      response = await LobbiesAPI.startQueue(userToken, lobby.id)
-    } else {
-      response = await LobbiesAPI.cancelQueue(userToken, lobby.id)
-    }
-
-    if (response.errorMsg) {
-      dispatch(
-        addToast({
-          content: response.errorMsg,
-          variant: 'error',
-        })
-      )
-    }
-  }
-
-  const handleCancelQueue = () => handleQueue('cancel')
-  const handleStartQueue = () => handleQueue('start')
 
   useEffect(() => {
     const lockIn = async () => {
@@ -113,14 +84,31 @@ export default function LobbyView() {
   }, [preMatch])
 
   return (
-    <Container className={style.container} column gap={40} align="center">
+    <Container
+      className={style.container}
+      column
+      gap={isLessThan2xl ? 30 : 40}
+      align="center"
+    >
       <Container className={style.header} gap={12} fitContent>
-        <Icon as={JoystickIcon} color="white" fontSize={30} />
+        <Icon
+          as={JoystickIcon}
+          color="white"
+          fontSize={{ base: 30, md: 28, '2xl': 30 }}
+        />
         <Container gap={8}>
-          <Text fontSize={22} fontWeight="light" textTransform="uppercase">
+          <Text
+            fontSize={{ base: 20, md: 18, '2xl': 20 }}
+            fontWeight="light"
+            textTransform="uppercase"
+          >
             Suba de nível e
           </Text>
-          <Text fontSize={22} fontWeight="semibold" textTransform="uppercase">
+          <Text
+            fontSize={{ base: 20, md: 18, '2xl': 20 }}
+            fontWeight="semibold"
+            textTransform="uppercase"
+          >
             fique entre os melhores
           </Text>
         </Container>
@@ -137,8 +125,10 @@ export default function LobbyView() {
           gap={12}
           align="center"
         >
-          <Text fontSize={18}>TDM 5x5</Text>
-          <Badge>Em breve</Badge>
+          <Text fontSize={{ base: 18, md: 16, '2xl': 18 }}>TDM 5x5</Text>
+          <Badge paddingBottom={{ base: '4px', md: '3px', '2xl': '4px' }}>
+            Em breve
+          </Badge>
         </Container>
 
         <Container
@@ -150,7 +140,7 @@ export default function LobbyView() {
           <Container className={style.arrowUpIcon} justify="center">
             <Icon as={ArrowUpFilledIcon} color="primary.400" />
           </Container>
-          <Text fontSize={18}>Ranqueada 5x5</Text>
+          <Text fontSize={{ base: 18, md: 16, '2xl': 18 }}>Ranqueada 5x5</Text>
         </Container>
 
         <Container
@@ -159,30 +149,20 @@ export default function LobbyView() {
           gap={12}
           align="center"
         >
-          <Text fontSize={18}>Personalizada</Text>
-          <Badge>Em breve</Badge>
+          <Text fontSize={{ base: 18, md: 16, '2xl': 18 }}>Personalizada</Text>
+          <Badge paddingBottom={{ base: '4px', md: '3px', '2xl': '4px' }}>
+            Em breve
+          </Badge>
         </Container>
       </Container>
 
       <Container className={style.lineup}>
         <LobbyLineup
           userPlayer={userPlayer}
-          isOwner={lobby.owner_id === user.id}
           otherPlayers={otherPlayers}
-          lobbyId={lobby.id}
-          queue={lobby.queue}
-        />
-      </Container>
-
-      <Container className={style.footer} fitContent>
-        <LobbyPlayButton
-          queueTime={lobby.queue && lobby.queue_time}
-          restrictionCountdown={lobby.restriction_countdown}
-          restricted={lobby.restriction_countdown}
-          disabled={(!isOwner && !lobby.queue) || preMatch || match}
-          onClick={
-            lobby.queue_time !== null ? handleCancelQueue : handleStartQueue
-          }
+          lobby={lobby}
+          preMatch={preMatch}
+          match={match}
         />
       </Container>